Extract retention jobs into src/core/retention.ts
Move validation scheduling, approved/non-approved retention reconcile,
history pruning, audit purge, and user-state validation into a
dedicated module. Typecheck, 167 tests, and build all pass.
